+++ Packages/best.crux.core/Runtime/Attributes/EndRevealAreaAttribute.cs
@@ -1,0 +1,12 @@
+using System;
+using JetBrains.Annotations;
+
+namespace Crux.Core.Runtime.Attributes
+{
+    [PublicAPI]
+    [AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
+    public class EndRevealAreaAttribute : Attribute
+    {
+        
+    }
+}